VIX-3697 Moved String Type to Color Wizard page
